--- Test slack
@@ -1,1 +1,0 @@
-test slack
