--- wct_app/.version
@@ -1,1 +1,0 @@
-1.2.3
