--- wct_app/data/widget.store
@@ -1,39 +1,0 @@
-
-
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-
-
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-
-
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-(ಠ_ಠ)foo(ಠ_ಠ)
-
-
-(ಠ_ಠ)foo(ಠ_ಠ)
