--- wct_app/main/__init__.py
@@ -1,1 +1,0 @@
-
