+++ widget.store
@@ -1,0 +1,12 @@
+
+
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
+(ಠ_ಠ)foo(ಠ_ಠ)
